refactor: reorganize content scripts and update import paths
- Replaced old content script paths in manifest.json for YouTube and Notion.
- Updated background.js to import new manager scripts for YouTube and Notion.
- Refactored options.js and actionButton.js to use updated utility paths.
- Removed deprecated common scripts and consolidated functionality into new modules.
- Cleaned up unused code and improved structure for better maintainability.
