+++ components/common/managerNotion.js
@@ -36,4 +36,4 @@ chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
   }
 });
 
-console.log('[background] notionContentIntercept initialized');
+console.log('[background] notion manager initialized');
