+++ components/contentScripts/llm/pasteFilesAsAttachments.js
